requirements.txt
gradio
pillow
transformers
numpy
scipy
phonemizer
